:art: Create Struct Files
:truck: Created QueueFamilyIndices.h
:truck: Created SwapChainSupportDetails.h
:truck: Created Vertex.h
