+++ core/narc_engine/include/QueueFamilyIndices.h
@@ -1,0 +1,17 @@
+#pragma once
+#include <cstdint>
+#include <optional>
+
+namespace NarcEngine
+{
+    struct QueueFamilyIndices
+    {
+        std::optional<uint32_t> GraphicsFamily;
+        std::optional<uint32_t> PresentFamily;
+
+        bool IsComplete() const
+        {
+            return GraphicsFamily.has_value() && PresentFamily.has_value();
+        }
+    };
+}
