+++ core/narc_engine/include/Vertex.h
@@ -1,0 +1,41 @@
+#pragma once
+
+#define GLFW_INCLUDE_VULKAN
+#include <array>
+#include <GLFW/glfw3.h>
+#include <glm/glm.hpp>
+
+namespace NarcEngine
+{
+    struct Vertex
+    {
+        glm::vec2 Pos;
+        glm::vec3 Color;
+
+        static VkVertexInputBindingDescription GetBindingDescription()
+        {
+            VkVertexInputBindingDescription bindingDescription{};
+            bindingDescription.binding = 0;
+            bindingDescription.stride = sizeof(Vertex);
+            bindingDescription.inputRate = VK_VERTEX_INPUT_RATE_VERTEX; //VERTEX or INSTANCE
+
+            return bindingDescription;
+        }
+
+        static std::array<VkVertexInputAttributeDescription, 2> GetAttributeDescriptions() //Fill attribute in vertx shader
+        {
+            std::array<VkVertexInputAttributeDescription, 2> attributeDescriptions{};
+            attributeDescriptions[0].binding = 0;
+            attributeDescriptions[0].location = 0;
+            attributeDescriptions[0].format = VK_FORMAT_R32G32_SFLOAT; //VEC2
+            attributeDescriptions[0].offset = offsetof(Vertex, Pos);
+
+            attributeDescriptions[1].binding = 0;
+            attributeDescriptions[1].location = 1;
+            attributeDescriptions[1].format = VK_FORMAT_R32G32B32_SFLOAT;
+            attributeDescriptions[1].offset = offsetof(Vertex, Color);
+
+            return attributeDescriptions;
+        }
+    };
+}
