+++ core/narc_engine/src/Application.cpp
